Renamed classes to avoid havign to use full path everywhere
